requirements.txt
asgiref==3.5.2
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==2.1.1
cryptography==38.0.4
Deprecated==1.2.13
dj-database-url==1.0.0
Django==4.1.4
django-cors-headers==3.7.0
django-heroku==0.3.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
gunicorn==20.1.0
idna==3.4
jwcrypto==1.4.2
oauthlib==3.2.2
psycopg2==2.9.5
pycparser==2.21
PyJWT==2.6.0
pytz==2022.6
requests==2.28.1
shortuuid==1.0.11
sqlparse==0.4.3
tzdata==2022.7
urllib3==1.26.13
whitenoise==6.2.0
wrapt==1.14.1